Rename virtual environment used in CI
Using the `circleci` tool allows us to build the project locally with
`circleci build --job=test`.

Unfortunately, the build command copies the local project workspace to
the Docker container instead of doing a fresh checkout.

This means that if a local environment is already installed with
`virtualenv env`, then it would cause unnecessary failures.

To better isolate these environments, the CI one has been renamed to
`env-ci`.

This feels unnecessary on one hand but logical on the other.