Add failing test for daylight savings time slots
Currently, we have a bug in production that during daylight savings,
members of the public can schedule appointments after one (1) hour.

The expected behaviour would be to be able to schedule after two (2)
hours.

Example:

When it's `10:30 AM` in London in summer time, the choice field contains
`11:30 AM` and `12:00 PM` as choices. The correct behaviour would be to
only display options from `12:30 PM`.